+++ packages/insomnia-inso/src/db/models/__tests__/api-spec.test.js
@@ -2,7 +2,7 @@
 import type { Database } from '../../index';
 import { emptyDb } from '../../index';
 import type { ApiSpec } from '../types';
-import { generateIdIsh } from '../utils';
+import { generateIdIsh } from '../util';
 import enquirer from 'enquirer';
 import { loadApiSpec, promptApiSpec } from '../api-spec';
