+++ packages/insomnia-inso/src/db/models/unit-test-suite.js
@@ -5,7 +5,7 @@ import { loadApiSpec } from './api-spec';
 import { mustFindSingleOrNone } from '../index';
 import { AutoComplete } from 'enquirer';
 import flattenDeep from 'lodash.flattendeep';
-import { generateIdIsh, getDbChoice, matchIdIsh } from './utils';
+import { generateIdIsh, getDbChoice, matchIdIsh } from './util';
 import { loadWorkspace } from './workspace';
 
 export const loadTestSuites = (db: Database, identifier: string): Array<UnitTestSuite> => {
